# apps/api/src/swagger.yaml
openapi: 3.0.3
info:
  title: AgriTrack Central API
  description: |
    AgriTrack Central Backend API for Smart CRM Machinery Monitoring.
    
    ## Features
    - **Machines**: Real-time machine monitoring and management
    - **Bookings**: Farmer booking system
    - **Analytics**: Dashboard statistics
    - **Notifications**: SMS, Push, and Email alerts
    - **Geofencing**: Location-based boundaries and breach detection
    - **Fuel Tracking**: Real-time fuel monitoring and theft detection
    - **Maintenance**: Scheduled maintenance management
    
    ## WebSocket Events
    Connect to `ws://localhost:3001` for real-time updates:
    - `device_update`: Real-time machine state
    - `booking_update`: Booking status changes
    - `initial_state`: All machines on connect
    
  version: 2.0.0
  contact:
    name: AgriTrack Team
    url: https://github.com/KD-3030/AgriTrack
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/v1/bookings:
    post:
      tags:
        - Bookings
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /api/v1/bookings/farmer/{id}:
    get:
      tags:
        - Bookings
      summary: Get farmer's bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
...
  /api/v1/farmers/{id}/bookings:
    get:
      tags:
        - Farmers
      summary: Get farmer's bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
        farmer_id:
          type: string
        machine_id:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
    BookingRequest:
      type: object
      required:
        - farmerId
        - machineId
        - startDate
        - endDate
      properties:
        farmerId:
          type: string
        machineId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
...